pyproject.toml
[project]
name = "mcp-sales-analytics"
version = "0.1.0"
description = "MCP server for sales analytics with forecasting capabilities"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.0.0",
    "pandas>=2.0.0",
    "prophet>=1.0.0",
    "google-genai>=0.1.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "ipython>=8.0.0",
    "jupyter>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
only-include = ["mcp_server.py", "mcp_client.py"]

[tool.uv]
dev-dependencies = []
